+++ src/Network/Network/LoadBalancer/FrontendIPConfiguration/NewAzureRmLoadBalancerFrontendIpConfigCommand.cs
@@ -116,6 +116,12 @@ namespace Microsoft.Azure.Commands.Network
         [ValidateNotNullOrEmpty]
         public string DdosCustomPolicyId { get; set; }
 
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Enables UDP flow tracking for traffic associated with the frontend IP configuration. When enabled, packets belonging to the same UDP flow are consistently directed to the same backend instance. This setting applies to all associated load balancing rules and takes precedence over rule-level connection tracking settings.",
+            ValueFromPipelineByPropertyName = true)]
+        public SwitchParameter EnableConnectionTracking { get; set; }
+
         public override void Execute()
         {
 
@@ -195,6 +201,8 @@ namespace Microsoft.Azure.Commands.Network
 
                 vFrontendIpConfigurations.DdosSettings.DdosCustomPolicy = new PSResourceId { Id = this.DdosCustomPolicyId };
             }
+
+            vFrontendIpConfigurations.EnableConnectionTracking = this.EnableConnectionTracking.IsPresent ? true : (bool?)null;
 
             var generatedId = string.Format(
                 "/subscriptions/{0}/resourceGroups/{1}/providers/Microsoft.Network/loadBalancers/{2}/{3}/{4}",